+++ yak_error_handler/lib/src/classes/all.dart
@@ -1,1 +1,2 @@
+export 'error.dart';
 export 'error_handler.dart';