+++ yak_error_handler/lib/src/provider/catch_error.dart
@@ -1,0 +1,5 @@
+import 'package:riverpod/all.dart';
+import 'package:yak_error_handler/src/typedef/catch.dart';
+import 'error_handler.dart';
+
+final catchError = Provider<Catch>((ref) => ref.read(errorHandler).call);