+++ yak_runner/lib/src/mixin/result/on_result_async/failure.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
+
 import '../../../classes/all.dart';
+import '../../../typedef/all.dart';
 
 mixin OnFailureAsync<S> on FailureBase<S> {
   Future<Result<T>> onResultAsync<T>(