+++ yak_runner/lib/src/runners/sync/nullary.dart
@@ -1,3 +1,4 @@
+import '../../typedef/all.dart';
 import '../../classes/all.dart';
 import '../../mixin/all.dart';
 import '../../delegates/all.dart';
@@ -6,7 +7,7 @@ class YakRunner<T> with YakRunnerTestMixin<T> implements Delegate<Result<T>> {
   const YakRunner(this.fun, [this.errorHandler])
       : assert(fun != null, ' a non null function must be provided');
   final T Function() fun;
-  final ErrorHandler errorHandler;
+  final Catch errorHandler;
 
   Result<T> call() {
     try {