+++ yak_runner/lib/src/runners/sync/unary.dart
@@ -1,7 +1,8 @@
+import '../../shared/all.dart';
+import '../../typedef/all.dart';
 import '../../classes/all.dart';
 import '../../mixin/all.dart';
 import '../../delegates/all.dart';
-import '../../shared/all.dart';
 
 class YakRunnerArg<T, S>
     with YakRunnerArgTestMixin<T, S>
@@ -10,7 +11,7 @@ class YakRunnerArg<T, S>
       : assert(fun != null, ' a non null function must be provided'),
         assert(S != typeVoid, '`S` must not be of type `void`');
   final T Function(S) fun;
-  final ErrorHandler errorHandler;
+  final Catch errorHandler;
 
   Result<T> call(S arg) {
     assert(arg != null, ' a non null `arg` must be provided');