+++ yak_runner/lib/src/typedef/catch.dart
@@ -1,0 +1,1 @@
+typedef void Catch(Object e, StackTrace s);