+++ yak_runner/lib/src/typedef/failure_of_type.dart
@@ -1,3 +1,3 @@
-import 'all.dart';
+import '../classes/all.dart';
 
 typedef Failure<T> FailureOfType<T>(Object e, StackTrace s);